when you click 'import' it renders the clusters yay
